Create 3D-Bifurc
Displays chaotic behavior of the logistic map in two ways.  See comments for details.
There's really no need to do this in 3D.  It was easier to use the free look camera for panning and zooming, and I thought there might be some interesting third dimension of data to plot by varying a parameter, but the sliders do it better.